+++ src/lib/ndpi_domains.c
@@ -93,7 +93,7 @@ int ndpi_load_domain_suffixes(struct ndpi_detection_module_struct *ndpi_str,
 
 const char* ndpi_get_host_domain_suffix(struct ndpi_detection_module_struct *ndpi_str,
 					const char *hostname,
-					u_int16_t *domain_id /* out */) {
+					u_int32_t *domain_id /* out */) {
   char *dot, *prev_dot;
 
   if(!ndpi_str || !hostname || !domain_id)
@@ -135,7 +135,7 @@ const char* ndpi_get_host_domain(struct ndpi_detection_module_struct *ndpi_str,
 				 const char *hostname) {
   const char *ret;
   char *dot, *first_dc;
-  u_int16_t domain_id, len;
+  u_int32_t domain_id, len;
   
   if(!ndpi_str || !hostname)
     return NULL;
